Refactor PU algorithm in a separate class, to allow alternative Puppi algorithms or implementations
