+++ L1Trigger/Phase2L1ParticleFlow/plugins/L1TPFProducer.cc
@@ -20,6 +20,7 @@
 #include "L1Trigger/Phase2L1ParticleFlow/interface/PFAlgoBase.h"
 #include "L1Trigger/Phase2L1ParticleFlow/interface/PFAlgo3.h"
 #include "L1Trigger/Phase2L1ParticleFlow/interface/PFAlgo2HGC.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/PuppiAlgo.h"
 
 //--------------------------------------------------------------------------------------------------
 class L1TPFProducer : public edm::stream::EDProducer<> {
@@ -33,7 +34,7 @@ class L1TPFProducer : public edm::stream::EDProducer<> {
         edm::EDGetTokenT<l1t::PFTrackCollection> tkCands_;
         float trkPt_, trkMaxChi2_;
         unsigned trkMinStubs_;
-        l1tpf_impl::PFAlgoBase::VertexAlgo vtxAlgo_;  
+        l1tpf_impl::PUAlgoBase::VertexAlgo vtxAlgo_;  
         edm::EDGetTokenT<std::vector<l1t::Vertex>> extVtx_;
 
         edm::EDGetTokenT<l1t::MuonBxCollection> muCands_;
@@ -45,6 +46,7 @@ class L1TPFProducer : public edm::stream::EDProducer<> {
 
         l1tpf_impl::RegionMapper l1regions_;
         std::unique_ptr<l1tpf_impl::PFAlgoBase> l1pfalgo_;
+        std::unique_ptr<l1tpf_impl::PUAlgoBase> l1pualgo_;
 
         // region of interest debugging
         float debugEta_, debugPhi_, debugR_;
@@ -67,6 +69,7 @@ L1TPFProducer::L1TPFProducer(const edm::ParameterSet& iConfig):
     hadPtCut_(iConfig.getParameter<double>("hadPtCut")),
     l1regions_(iConfig),
     l1pfalgo_(nullptr),
+    l1pualgo_(nullptr),
     debugEta_(iConfig.getUntrackedParameter<double>("debugEta",0)),
     debugPhi_(iConfig.getUntrackedParameter<double>("debugPhi",0)),
     debugR_(iConfig.getUntrackedParameter<double>("debugR",-1))
@@ -83,8 +86,6 @@ L1TPFProducer::L1TPFProducer(const edm::ParameterSet& iConfig):
     produces<l1t::PFCandidateCollection>("TKVtx");
 
     produces<float>("z0");
-    produces<float>("alphaCMed"); produces<float>("alphaCRms"); produces<float>("alphaFMed"); produces<float>("alphaFRms");
-
 
     for (auto & tag : iConfig.getParameter<std::vector<edm::InputTag>>("emClusters")) {
         emCands_.push_back(consumes<l1t::PFClusterCollection>(tag));
@@ -102,14 +103,23 @@ L1TPFProducer::L1TPFProducer(const edm::ParameterSet& iConfig):
         throw cms::Exception("Configuration", "FIXME: recover Bitwise PF");
     } else throw cms::Exception("Configuration", "Unsupported PFAlgo");
 
+    const std::string & pualgo = iConfig.getParameter<std::string>("puAlgo");
+    if (pualgo == "Puppi") {
+        l1pualgo_.reset(new l1tpf_impl::PuppiAlgo(iConfig));
+    } else throw cms::Exception("Configuration", "Unsupported PUAlgo");
+
     std::string vtxAlgo = iConfig.getParameter<std::string>("vtxAlgo");
-    if      (vtxAlgo == "TP")  vtxAlgo_ = l1tpf_impl::PFAlgoBase::TPVtxAlgo;
-    else if (vtxAlgo == "old") vtxAlgo_ = l1tpf_impl::PFAlgoBase::OldVtxAlgo;
+    if      (vtxAlgo == "TP")  vtxAlgo_ = l1tpf_impl::PUAlgoBase::TPVtxAlgo;
+    else if (vtxAlgo == "old") vtxAlgo_ = l1tpf_impl::PUAlgoBase::OldVtxAlgo;
     else if (vtxAlgo == "external") {
-        vtxAlgo_ = l1tpf_impl::PFAlgoBase::ExternalVtxAlgo;
+        vtxAlgo_ = l1tpf_impl::PUAlgoBase::ExternalVtxAlgo;
         extVtx_  = consumes<std::vector<l1t::Vertex>>(iConfig.getParameter<edm::InputTag>("vtxCollection"));
     } else throw cms::Exception("Configuration") << "Unsupported vtxAlgo " << vtxAlgo << "\n";
 
+
+    for (const std::string & label : l1pualgo_->puGlobalNames()) {
+        produces<float>(label); 
+    }
 }
 
 // ------------ method called to produce the data  ------------
@@ -168,7 +178,7 @@ L1TPFProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
 
     // Then do the vertexing, and save it out
     float z0;
-    if (vtxAlgo_ == l1tpf_impl::PFAlgoBase::ExternalVtxAlgo) {
+    if (vtxAlgo_ == l1tpf_impl::PUAlgoBase::ExternalVtxAlgo) {
         edm::Handle<std::vector<l1t::Vertex>> vtxHandle;
         iEvent.getByToken(extVtx_, vtxHandle);
         z0 = 0;
@@ -179,7 +189,7 @@ L1TPFProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
             if (myptsum > ptsum) { z0 = vtx.z0(); ptsum = myptsum; }
         }
     }
-    l1pfalgo_->doVertexing(l1regions_.regions(), vtxAlgo_, z0);
+    l1pualgo_->doVertexing(l1regions_.regions(), vtxAlgo_, z0);
     iEvent.put(std::make_unique<float>(z0), "z0");
 
     // Then also save the tracks with a vertex cut
@@ -188,20 +198,23 @@ L1TPFProducer::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
     // Then run PF in each region
     for (auto & l1region : l1regions_.regions()) {
         l1pfalgo_->runPF(l1region);
-        l1pfalgo_->runChargedPV(l1region, z0);
+        l1pualgo_->runChargedPV(l1region, z0);
     }
     // save PF into the event
     iEvent.put(l1regions_.fetch(false), "PF");
 
     // Then get our alphas (globally)
-    float alphaCMed, alphaCRms, alphaFMed, alphaFRms;
-    l1pfalgo_->computePuppiMedRMS(l1regions_.regions(), alphaCMed, alphaCRms, alphaFMed, alphaFRms);
-    iEvent.put(std::make_unique<float>(alphaCMed), "alphaCMed"); iEvent.put(std::make_unique<float>(alphaCRms), "alphaCRms");
-    iEvent.put(std::make_unique<float>(alphaFMed), "alphaFMed"); iEvent.put(std::make_unique<float>(alphaFRms), "alphaFRms");
+    std::vector<float> puGlobals;
+    l1pualgo_->doPUGlobals(l1regions_.regions(), -1., puGlobals); // FIXME we don't have yet an external PU estimate
+    const std::vector<std::string> & puGlobalNames = l1pualgo_->puGlobalNames();
+    assert(puGlobals.size() == puGlobalNames.size());
+    for (unsigned int i = 0, n = puGlobalNames.size(); i < n; ++i) {
+        iEvent.put(std::make_unique<float>(puGlobals[i]), puGlobalNames[i]);
+    }
 
     // Then run puppi (regionally)
     for (auto & l1region : l1regions_.regions()) {
-        l1pfalgo_->runPuppi(l1region, -1., alphaCMed, alphaCRms, alphaFMed, alphaFRms);
+        l1pualgo_->runNeutralsPU(l1region, -1., puGlobals);
     }
     // and save puppi
     iEvent.put(l1regions_.fetch(true), "Puppi");
